refactor(cli): split flat image.rs + catalog.rs into directories (Plan 153)
The two remaining flat multi-subaction files become per-subaction
directories (Plan 153, folded into Plan 178):

- catalog.rs → catalog/{mod,list,search,info}.rs
- image.rs → image/{mod,pull,ls,inspect,rm}.rs; the shared OCI machinery
  (helpers + provenance types) stays in mod.rs as pub(super), the four
  thin subaction handlers move to siblings

Move-only — no behavior change. mvm-cli 941/941; audit coverage gate green;
clippy -D warnings clean.
